Gerneral re-work: It turns out using bits alone does not do the trick. Do have piTest proceeding as proposal and work at least at this direction
Clean-ed up a lot
Provided shortcut for 8bit-reads
Moved piControlFunc for the sake of better survey
Changed return type of reads to long in order to have the 32bit reads supported and negative return values for errors too

+++ src/org/clehne/revpi/ledio/LedInOut.java
@@ -43,11 +43,11 @@ public final class LedInOut extends LibLedInOut implements Closeable {
 	}
 	
 	public int getValue() throws IOException {
-		int val = _getValue(_fd);
+		long val = _getValue(_fd);
 		if(val < 0) {
 			throw new IOException("Error getting value, retCode " + val);
 		}
-		return val;
+		return Long.valueOf(val).intValue();
 	}
 	
 
@@ -55,7 +55,7 @@ public final class LedInOut extends LibLedInOut implements Closeable {
 	private static native void _close(final int fd) throws IOException;
 
 	private static native int _setValue(final int fd, int value);
-	private static native int _getValue(final int fd);
+	private static native long _getValue(final int fd);
 	private static native int _setBit(final int fd, int bit, int value);
 	private static native int _getBit(final int fd, int bit);
 }
